Moves AboutContent into 'src/components'
